requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
python-jose==3.3.0
python-docx==0.8.11
matplotlib==3.7.1
numpy==1.24.3 